+++ CHANGELOG.rst
@@ -9,13 +9,14 @@ v1.1.0 (unreleased)
 The 1.1.0 release is focused on new features to improve ease of use.
 
 * Provide a button to toggle line wrapping, contributed by Kwpolska_.
+* Auto-delete pastes on view when they've expired. (#63)
 
 v1.0.1 (20200326)
 *****************
 
 A quick bugfix release to depend on a newer version of ``pygments-better-html``.
 
-* #67_: Empty lines don't survive copy/paste
+* Empty lines don't survive copy/paste. (#67)
 
 v1.0.0 (20200323)
 *****************
@@ -24,4 +25,3 @@ After a period of darkness (changelog-wise) version 1.0.0 was released and this
 changelog created.
 
 .. _Kwpolska: https://github.com/Kwpolska
-.. _67: https://github.com/supakeen/pinnwand/issues/67